Move Client conditions code into client package
