+++ pkg/controller/client/conditions.go
@@ -14,25 +14,21 @@
 // limitations under the License.
 //
 
-package util
+package client
 
 import (
 	"fmt"
 
-	securityv1 "github.com/IBM/ibm-iam-operator/pkg/apis/oidc/v1"
+	oidcv1 "github.com/IBM/ibm-iam-operator/pkg/apis/oidc/v1"
 	metav1 "k8s.io/apimachinery/pkg/apis/meta/v1"
-	"k8s.io/apimachinery/pkg/util/clock"
 )
-
-// Clock is defined as a package var so it can be stubbed out during tests.
-var Clock clock.Clock = clock.RealClock{}
 
 // ClientHasCondition will return true if the given Client has a
 // condition matching the provided ClientCondition.
 // Only the Type and Status field will be used in the comparison, meaning that
 // this function will return 'true' even if the Reason, Message and
 // LastTransitionTime fields do not match.
-func ClientHasCondition(client *securityv1.Client, c securityv1.ClientCondition) bool {
+func ClientHasCondition(client *oidcv1.Client, c oidcv1.ClientCondition) bool {
 	if client == nil {
 		return false
 	}
@@ -53,8 +49,8 @@ func ClientHasCondition(client *securityv1.Client, c securityv1.ClientCondition)
 // - If a condition of the same type and different state already exists, the
 //   condition will be updated and the LastTransitionTime set to the current
 //   time.
-func SetClientCondition(client *securityv1.Client, conditionType securityv1.ClientConditionType, status securityv1.ConditionStatus, reason, message string) {
-	newCondition := securityv1.ClientCondition{
+func SetClientCondition(client *oidcv1.Client, conditionType oidcv1.ClientConditionType, status oidcv1.ConditionStatus, reason, message string) {
+	newCondition := oidcv1.ClientCondition{
 		Type:    conditionType,
 		Status:  status,
 		Reason:  reason,
